metrics and specs clients package refactoring
