+++ vertical-pod-autoscaler/recommender/input/history/history_provider_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package signals
+package history
 
 import (
 	"fmt"
