+++ vertical-pod-autoscaler/recommender/input/history/prometheus_client.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package signals
+package history
 
 import (
 	"fmt"
