+++ vertical-pod-autoscaler/recommender/input/history/prometheus_util_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package signals
+package history
 
 import (
 	"strings"
