+++ vertical-pod-autoscaler/recommender/input/spec/spec_client.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package cluster
+package spec
 
 import (
 	"k8s.io/api/core/v1"
