+++ vertical-pod-autoscaler/recommender/input/spec/spec_client_test_util.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package cluster
+package spec
 
 import (
 	"fmt"
