+++ vertical-pod-autoscaler/recommender/main.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/golang/glog"
 	kube_flag "k8s.io/apiserver/pkg/util/flag"
-	"k8s.io/autoscaler/vertical-pod-autoscaler/recommender/signals"
+	"k8s.io/autoscaler/vertical-pod-autoscaler/recommender/input/history"
 	"k8s.io/client-go/rest"
 	kube_restclient "k8s.io/client-go/rest"
 )
@@ -37,7 +37,7 @@ func main() {
 	glog.Infof("Running VPA Recommender")
 
 	config := createKubeConfig()
-	recommender := NewRecommender(config, *metricsFetcherInterval, signals.NewPrometheusHistoryProvider(*prometheusAddress))
+	recommender := NewRecommender(config, *metricsFetcherInterval, history.NewPrometheusHistoryProvider(*prometheusAddress))
 	recommender.Run()
 }
